requirements.txt
absl-py==2.1.0
accelerate==1.5.2
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astroid==2.5
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==24.2.0
Automat==24.8.1
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
bibtexparser==1.4.2
bitsandbytes==0.45.4
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
constantly==23.10.4
contourpy==1.3.1
crossref-commons==0.0.7
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
distro==1.9.0
docutils==0.21.2
durationpy==0.9
einops==0.8.1
exceptiongroup==1.2.2
executing==2.1.0
facexlib==0.3.0
faiss-cpu==1.10.0
fake-useragent==2.1.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
filterpy==1.4.5
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.55.0
fqdn==1.5.1
free_proxy==1.1.3
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
html5lib==1.1
HTMLParser==0.0.2
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hyperlink==21.0.0
idna==2.10
imageio==2.36.0
imagesize==1.4.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
incremental==24.7.2
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
isort==5.13.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kubernetes==32.0.1
langchain==0.3.21
langchain-chroma==0.2.2
langchain-community==0.3.20
langchain-core==0.3.49
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.7
langsmith==0.3.19
lazy-object-proxy==1.10.0
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mccabe==0.6.1
mdurl==0.1.2
mistune==3.1.3
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.2.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.21.0
open3d==0.18.0
opencv-python==4.10.0.84
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
posthog==3.23.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.1
Protego==0.4.0
protobuf==5.29.4
proxy.py==2.4.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyChainedProxy==1.3
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
PyDispatcher==2.0.7
Pygments==2.18.0
pylint==2.7.1
pyOpenSSL==25.0.0
-e git+https://github.com/ferru97/PyPaperBot.git@a65bf7071478d79ee312d7d21154aad4c806e198#egg=PyPaperBot
pyparsing==3.1.4
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.3.0
queuelib==1.7.0
ratelimit==2.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scholarly==1.7.11
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
selenium==4.25.0
Send2Trash==1.8.3
sentence-transformers==4.0.1
sentencepiece==0.2.0
service-identity==24.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.4.0
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
trio==0.26.2
trio-websocket==0.11.1
triton==3.2.0
Twisted==24.11.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
uri-template==1.3.0
urllib3==2.2.3
uuid==1.30
uvicorn==0.34.0
uvloop==0.21.0
w3lib==2.3.1
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.12.1
wsproto==1.2.0
XlsxWriter==3.2.2
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
